agregando endpoint

+++ src/main/java/com/example/Sistema/para/Bazar/model/Producto.java
@@ -16,12 +16,12 @@ public class Producto {
     private String nombre;
     private String marca;
     private double costo;
-    private double cantidad_disponible;
+    private int cantidad_disponible;
 
     public Producto() {
     }
 
-    public Producto(Long codigo_producto, String nombre, String marca, double costo, double cantidad_disponible) {
+    public Producto(Long codigo_producto, String nombre, String marca, double costo, int cantidad_disponible) {
         this.codigo_producto = codigo_producto;
         this.nombre = nombre;
         this.marca = marca;
